♻️ Home > Counter > create counter component
